+++ src/main/java/com/alevel/nix7/adminassistant/service/impl/ProcedureServiceImpl.java
@@ -1,5 +1,6 @@
 package com.alevel.nix7.adminassistant.service.impl;
 
+import com.alevel.nix7.adminassistant.exceptions.AssistantException;
 import com.alevel.nix7.adminassistant.model.procedure.Procedure;
 import com.alevel.nix7.adminassistant.model.procedure.ProcedureRequest;
 import com.alevel.nix7.adminassistant.model.procedure.ProcedureResponse;
@@ -15,7 +16,9 @@ public class ProcedureServiceImpl implements ProcedureService {
 
     private final ProcedureRepository procedureRepository;
     private final SpecialistRepository specialistRepository;
-    private final long MINUTE = 60_000;
+    private static final long MINUTE = 60_000;
+    private static final long MIN_DURATION = 10;
+    private static final long MAX_DURATION = 240;
 
     public ProcedureServiceImpl(ProcedureRepository procedureRepository,
                                 SpecialistRepository specialistRepository) {
@@ -25,32 +28,28 @@ public class ProcedureServiceImpl implements ProcedureService {
 
     @Override
     public ProcedureResponse create(ProcedureRequest procedure, Long idSpecialist) {
+        checkProcedure(procedure);
         return save(procedure, idSpecialist);
-    }
-
-    private ProcedureResponse save(ProcedureRequest request, Long idSpecialist) {
-        var procedure = new Procedure();
-        procedure.setName(request.name());
-        procedure.setDuration(request.duration() * MINUTE);
-        procedure.setPrice(request.price());
-        procedure.setSpecialist(specialistRepository.getById(idSpecialist));
-        procedureRepository.save(procedure);
-        return ProcedureResponse.fromProcedure(procedure);
     }
 
     @Override
     public void delete(Long id) {
+        if (!procedureRepository.existsById(id)) {
+            throw AssistantException.procedureNotFound(id);
+        }
         procedureRepository.deleteById(id);
     }
 
     @Override
     public ProcedureResponse getById(Long id) {
-        return ProcedureResponse.fromProcedure(procedureRepository.findById(id).orElseThrow());
+        return ProcedureResponse.fromProcedure(procedureRepository.findById(id)
+                .orElseThrow(() -> AssistantException.procedureNotFound(id)));
     }
 
     @Override
     public List<ProcedureResponse> findAllBySpecialist(Long id) {
-        return procedureRepository.findAllBySpecialist(specialistRepository.getById(id))
+        return procedureRepository.findAllBySpecialist(specialistRepository.findById(id)
+                        .orElseThrow(() -> AssistantException.workerNotFound(id)))
                 .stream().map(ProcedureResponse::fromProcedure).toList();
     }
 
@@ -59,5 +58,25 @@ public class ProcedureServiceImpl implements ProcedureService {
         return procedureRepository.findAll().stream()
                 .map(ProcedureResponse::fromProcedure)
                 .toList();
+    }
+
+    private void checkProcedure(ProcedureRequest procedure) {
+        if (procedure.duration() < MIN_DURATION) {
+            throw AssistantException.invalidDuration("Too short duration");
+        }
+        if (procedure.duration() > MAX_DURATION) {
+            throw AssistantException.invalidDuration("Too long duration");
+        }
+    }
+
+    private ProcedureResponse save(ProcedureRequest request, Long idSpecialist) {
+        var procedure = new Procedure();
+        procedure.setName(request.name());
+        procedure.setDuration(request.duration() * MINUTE);
+        procedure.setPrice(request.price());
+        procedure.setSpecialist(specialistRepository.findById(idSpecialist)
+                .orElseThrow(() -> AssistantException.workerNotFound(idSpecialist)));
+        procedureRepository.save(procedure);
+        return ProcedureResponse.fromProcedure(procedure);
     }
 }
